Fix Loop Bug
[Problem]
Bot is starting but stuck in a startup mode?

[Solution]
Looks like we have to specify not to show any output and to log
properly.

+++ scripts/stopbot.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-echo Killing bot...
-kill -9 $(<"/home/ec2-user/bot.pid")
-rm /home/ec2-user/bot.pid
+bot_pid_file=/home/ec2-user/bot.pid
+if test -f "$bot_pid_file"; then
+    echo Killing bot...
+    kill -9 $(<"$bot_pid_file")
+    rm /home/ec2-user/bot.pid
+else
+    echo No bot to kill!
+fi
